Reorganize testing materials
