+++ tests/csvs/input.csv
@@ -1,0 +1,2 @@
+Destination,NumericLabels
+tests/audio/veryshort.wav,0
