--- tests/input.csv
@@ -1,2 +1,0 @@
-Destination,NumericLabels
-tests/veryshort.wav,0
